requirements.txt
Django==4.2.7
djongo==1.3.6
pymongo==4.6.1
python-decouple==3.8
requests==2.31.0
djangorestframework==3.14.0
